+++ OsuPlayer/Modules/Audio/Player.cs
@@ -12,6 +12,7 @@ using OsuPlayer.Extensions;
 using OsuPlayer.IO.Storage.Blacklist;
 using OsuPlayer.IO.Storage.Playlists;
 using OsuPlayer.Modules.Audio.Engine;
+using OsuPlayer.Modules.Services;
 using OsuPlayer.Network.Discord;
 
 namespace OsuPlayer.Modules.Audio;
